pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-knowledge-graph"
version = "0.1.0"
description = "Personal Knowledge Graph RAG + Mind-Map Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rag", "knowledge-graph", "ai", "llm", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.13",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sse-starlette>=1.8.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "neo4j>=5.14.0",
    "graphiti-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "watchdog>=3.0.0",
    "filetype>=1.2.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",
    "langchain>=0.3.1",
    "langchain-community>=0.3.1",
    "langchain-core>=0.3.1",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=3.17.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.5.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
rag-cli = "src.cli:main"
rag-ingest = "src.ingestion.cli:main"
rag-api = "src.api.main:run"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
